Split to own model
