+++ OpenKh.Command.AnbMaker/Utils/Builder/Models/AQuaternionKey.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OpenKh.Command.AnbMaker.Utils.Builder.Models
+{
+    public class AQuaternionKey
+    {
+        public float Time { get; set; }
+        public Quaternion Value { get; set; }
+    }
+}
