+++ OpenKh.Command.AnbMaker/Utils/Builder/Models/AVectorKey.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OpenKh.Command.AnbMaker.Utils.Builder.Models
+{
+    public class AVectorKey
+    {
+        public float Time { get; set; }
+        public Vector3 Value { get; set; }
+    }
+}
